+++ sdk/appconfiguration/azure-data-appconfiguration/src/test/java/com/azure/data/appconfiguration/ConfigurationAsyncClientTest.java
@@ -726,7 +726,6 @@ public class ConfigurationAsyncClientTest extends ConfigurationClientTestBase {
                 .verifyComplete();
     }
 
-    @Disabled("reactor netty aggregate method aggregate random order of context. Tracking issue #7771")
     /**
      * Verifies that, given a ton of revisions, we can list the revisions ConfigurationSettings using pagination
      * (ie. where 'nextLink' has a URL pointing to the next page of results.)
@@ -752,7 +751,6 @@ public class ConfigurationAsyncClientTest extends ConfigurationClientTestBase {
             .verifyComplete();
     }
 
-    @Disabled("reactor netty aggregate method aggregate random order of context. Tracking issue #7771")
     /**
      * Verifies that, given a ton of revisions, we can list the revisions ConfigurationSettings using pagination and stream is invoked multiple times.
      * (ie. where 'nextLink' has a URL pointing to the next page of results.)
@@ -783,7 +781,6 @@ public class ConfigurationAsyncClientTest extends ConfigurationClientTestBase {
         assertEquals(numberExpected, configurationSettingList2.size());
     }
 
-    @Disabled("reactor netty aggregate method aggregate random order of context. Tracking issue #7771")
     /**
      * Verifies that, given a ton of revisions, we can list the revisions ConfigurationSettings using pagination and stream is invoked multiple times.
      * (ie. where 'nextLink' has a URL pointing to the next page of results.)
@@ -814,7 +811,6 @@ public class ConfigurationAsyncClientTest extends ConfigurationClientTestBase {
         assertEquals(numberExpected, configurationSettingList2.size());
     }
 
-    @Disabled("reactor netty aggregate method aggregate random order of context. Tracking issue #7771")
     /**
      * Verifies that, given a ton of existing settings, we can list the ConfigurationSettings using pagination
      * (ie. where 'nextLink' has a URL pointing to the next page of results.