+++ sdk/storage/azure-storage-blob/CHANGELOG.md
@@ -1,6 +1,7 @@
 # Release History
 
 ## 12.4.0-beta.1 (Unreleased)
+- Added ability to access BlobProperties from BlobInputStream.
 - Modified downloadToFile to populate BlobProperties.blobSize to be the actual blob size instead of the content length of the first range. 
 
 ## 12.3.0 (2020-01-16)